requirements.txt
pandas
laspy[lazrs]
scikit-learn
tensorboard==2.11.2
tqdm
open3d
matplotlib>=2.2.3
vispy>=0.5.3
PyYAML>=5.1.1
imgui[glfw]>=1.0.0 
glfw>=1.8.3
PyOpenGL>=3.1.0
pyqt5>=5.8.1.1
